# docs/openapi.yaml
openapi: 3.0.3
info:
  title: jjobb_v2 Graduate Network API
  version: 2.0.0-sprint4
  description: Sprint 4 recommendations, field trips, networking school scope (+ Sprint 2–3). v1 `/api` keys are preserved.
servers:
  - url: http://localhost:5000
    description: Local / Docker Compose backend
...
paths:
...
  /api/field-trips:
    get:
      tags: [FieldTrips]
      summary: List field trips (school scoped, REQ-TRP-001)
      responses:
        '200': { description: '{ field_trips }' }
    post:
      tags: [FieldTrips]
      security: [{ bearerAuth: [] }]
      summary: Create field trip
      responses:
        '201': { description: created }
  /api/field-trips/{id}/apply:
    post:
      tags: [FieldTrips]
      security: [{ bearerAuth: [] }]
      summary: Apply with capacity/fifo|approval (REQ-TRP-002)
      responses:
        '201': { description: application }
        '400': { description: CAPACITY }
        '403': { description: cross-school }
  /api/field-trips/{id}/roster:
    get:
      tags: [FieldTrips]
      security: [{ bearerAuth: [] }]
      summary: Roster (REQ-TRP-003)
      responses:
        '200': { description: roster }
        '403': { description: cross-school }
  /api/field-trips/{id}/report:
    get:
      tags: [FieldTrips]
      security: [{ bearerAuth: [] }]
      summary: After-report (REQ-TRP-003)
      responses:
        '200': { description: '{ trip, report }' }
        '403': { description: cross-school }
    put:
      tags: [FieldTrips]
      security: [{ bearerAuth: [] }]
      summary: Upsert after-report (REQ-TRP-003)
      responses:
        '200': { description: saved }
        '403': { description: cross-school }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT